fix: Update CautionCard Component
