+++ src/components/molecules/CautionCard/type.ts
@@ -1,0 +1,8 @@
+export type CautionType = 'info' | 'error' | 'warning'
+
+export interface Props {
+  /**
+   * タイプ
+   */
+  type: CautionType
+}
